Bastardized BlockFace implementation
